+++ packages/mui/src/types.ts
@@ -8,9 +8,14 @@
 // See: https://www.typescriptlang.org/docs/handbook/declaration-merging.html#module-augmentation
 
 import type { RoleProps } from "@ariakit/react/role";
+import type { AlertProps } from "@mui/material/Alert";
 import type { BadgeProps } from "@mui/material/Badge";
 import type { IconButtonProps } from "@mui/material/IconButton";
-import type { CommonProps } from "@mui/material/OverridableComponent";
+import type {
+	CommonProps,
+	DefaultComponentProps,
+	OverridableTypeMap,
+} from "@mui/material/OverridableComponent";
 import type { TabProps } from "@mui/material/Tab";
 import type { TableCellProps as MuiTableCellProps } from "@mui/material/TableCell";
 import type { TabsProps } from "@mui/material/Tabs";
@@ -19,7 +24,10 @@ import type {
 	TextFieldVariants,
 } from "@mui/material/TextField";
 import type { TooltipProps } from "@mui/material/Tooltip";
-import type { TypographyProps } from "@mui/material/Typography";
+import type {
+	TypographyProps,
+	TypographyTypeMap,
+} from "@mui/material/Typography";
 import type * as React from "react";
 
 declare module "@mui/material/OverridableComponent" {
@@ -39,6 +47,15 @@ declare module "@mui/material/OverridableComponent" {
 
 		/** @deprecated Use `render` prop instead. */
 		component?: React.ElementType;
+	}
+
+	interface OverridableComponent<TypeMap extends OverridableTypeMap> {
+		// biome-ignore lint/style/useShorthandFunctionType: Interface with call signature is necessary when merging.
+		(
+			props:
+				| DefaultComponentProps<TypeMap>
+				| TypographyOverridableComponentProps<TypeMap>,
+		): React.JSX.Element | null;
 	}
 }
 
@@ -62,13 +79,24 @@ declare module "@mui/material/Alert" {
 		standard: false;
 	}
 
+	interface AlertPropsColorOverrides {
+		none: true;
+	}
+
 	interface AlertOwnProps {
 		/**
 		 * The default variant with `@stratakit/mui` is `"outlined"`.
 		 *
 		 * @default 'outlined'
 		 */
-		variant?: "filled" | "outlined";
+		variant?: AlertProps["variant"];
+
+		/**
+		 * The default severity with `@stratakit/mui` is `"none"`.
+		 *
+		 * @default 'none'
+		 */
+		severity?: AlertProps["severity"];
 	}
 }
 
@@ -150,6 +178,17 @@ declare module "@mui/material/Button" {
 declare module "@mui/material/BottomNavigationAction" {
 	interface BottomNavigationActionOwnProps {
 		LinkComponent?: never;
+	}
+}
+
+declare module "@mui/material/Card" {
+	interface CardOwnProps {
+		/**
+		 * The default variant with `@stratakit/mui` is `"outlined"`.
+		 *
+		 * @default 'outlined'
+		 */
+		variant?: "outlined" | "elevation";
 	}
 }
 
@@ -374,8 +413,21 @@ declare module "@mui/material/Tabs" {
 	}
 
 	interface TabsOwnProps {
+		/**
+		 * The size of the tab buttons.
+		 *
+		 * @default 'medium'
+		 */
+		size?: "small" | "medium";
+
 		/** @deprecated DO NOT USE */
 		indicatorColor?: TabsProps["indicatorColor"];
+
+		/** @deprecated DO NOT USE. */
+		allowScrollButtonsMobile?: boolean;
+
+		/** @deprecated DO NOT USE. */
+		scrollButtons?: TabsProps["scrollButtons"];
 	}
 }
 
@@ -441,17 +493,84 @@ declare module "@mui/material/Tooltip" {
 	}
 }
 
+// These headings variants are declared separately from TypographyPropsVariantOverrides,
+// so that we can force the `render` prop to be required for these variants.
+type TypographyHeadingVariantProps = {
+	variant:
+		| "display-lg"
+		| "display-md"
+		| "display-sm"
+		| "headline-lg"
+		| "headline-md"
+		| "headline-sm"
+		| "subtitle-lg"
+		| "subtitle-md"
+		| "subtitle-sm"
+		| "h1"
+		| "h2"
+		| "h3"
+		| "h4"
+		| "h5"
+		| "h6"
+		| "subtitle1"
+		| "subtitle2";
+	/**
+	 * When using a heading-like `variant`, the `render` prop must be manually set to the most semantically appropriate element.
+	 *
+	 * Pick the most appropriate heading element ([`<h1>` to `<h6>`](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/Heading_Elements))
+	 * required to maintain proper [heading structure](https://www.a11yproject.com/posts/how-to-accessible-heading-structure/) in your application.
+	 *
+	 * @example
+	 * ```tsx
+	 * render={<h2 />}>
+	 * ```
+	 *
+	 * Do not use heading elements when you simply want to grab attention with large variants.
+	 */
+	render: NonNullable<RoleProps["render"]>;
+};
+
+type TypographyOverridableComponentProps<TypeMap extends OverridableTypeMap> =
+	TypeMap extends TypographyTypeMap
+		? Omit<
+				DefaultComponentProps<TypeMap>,
+				keyof TypographyHeadingVariantProps
+			> &
+				TypographyHeadingVariantProps
+		: never;
+
 declare module "@mui/material/Typography" {
 	interface TypographyPropsColorOverrides {
 		secondary: false;
 		textTertiary: true;
 	}
 
+	interface TypographyPropsVariantOverrides {
+		// Additional custom variants (non-heading).
+		"body-lg": true;
+		"body-md": true;
+		"body-sm": true;
+		"caption-lg": true;
+		"caption-md": true;
+		"caption-sm": true;
+		"mono-sm": true;
+
+		// Stock MUI heading variants are removed here and re-added above, with the `render` prop required.
+		h1: false;
+		h2: false;
+		h3: false;
+		h4: false;
+		h5: false;
+		h6: false;
+		subtitle1: false;
+		subtitle2: false;
+	}
+
 	interface TypographyOwnProps {
 		/**
-		 * The default variant with `@stratakit/mui` is `"body2"`.
+		 * The default variant with `@stratakit/mui` is `"inherit"`.
 		 *
-		 * @default "body2"
+		 * @default "inherit"
 		 */
 		variant?: TypographyProps["variant"];
 	}