+++ packages/mui/src/~createTheme.tsx
@@ -21,7 +21,7 @@ import {
 	MuiAutocompleteClearIndicator,
 } from "./~components/MuiAutocomplete.js";
 import { MuiAvatarGroup } from "./~components/MuiAvatarGroup.js";
-import { MuiBadge } from "./~components/MuiBadge.js";
+import { MuiBadge, MuiBadgeBadge } from "./~components/MuiBadge.js";
 import { MuiBottomNavigationAction } from "./~components/MuiBottomNavigation.js";
 import { MuiButtonBase } from "./~components/MuiButtonBase.js";
 import {
@@ -47,8 +47,9 @@ import {
 	MuiTableCell,
 	MuiTableHead,
 } from "./~components/MuiTable.js";
+import { MuiTab, MuiTabs } from "./~components/MuiTabs.js";
 import { MuiToggleButton } from "./~components/MuiToggleButton.js";
-import { MuiTypography } from "./~components/MuiTypography.js";
+import { MuiTypography, variantMapping } from "./~components/MuiTypography.js";
 import {
 	ArrowDownIcon,
 	CaretsUpDownIcon,
@@ -67,8 +68,14 @@ import {
 import type { RoleProps } from "@ariakit/react/role";
 import type { ColorSystemOptions } from "@mui/material/styles";
 
+interface CreateThemeArgs {
+	portalContainer?: HTMLElement | null;
+}
+
 /** Creates a StrataKit theme for MUI. Should be used with MUI's `ThemeProvider`. */
-function createTheme() {
+function createTheme(args: CreateThemeArgs) {
+	const { portalContainer: container } = args;
+
 	// Map the JS palette back to MUI's own CSS variables, which will then be mapped to the correct StrataKit tokens in CSS.
 	// (This is a fallback for any code that uses MUI's theme.palette values directly instead of CSS variables)
 	const palette = {
@@ -106,15 +113,19 @@ function createTheme() {
 			// These are only hardcoded here as fallback. The CSS will take precedence.
 			body1: { fontSize: 16 },
 			body2: { fontSize: 14 },
-			h1: { fontSize: 48 },
-			h2: { fontSize: 40 },
-			h3: { fontSize: 32 },
-			h4: { fontSize: 28 },
-			h5: { fontSize: 24 },
-			h6: { fontSize: 20 },
+			h1: { fontSize: 20 },
+			h2: { fontSize: 18 },
+			h3: { fontSize: 16 },
+			h4: { fontSize: 14 },
+			h5: { fontSize: 12 },
+			h6: { fontSize: 12 },
 			caption: { fontSize: 12 },
-			subtitle1: { fontSize: 12 },
-			subtitle2: { fontSize: 11 },
+			overline: {
+				fontSize: 12,
+				textTransform: "none",
+			},
+			subtitle1: { fontSize: 16 },
+			subtitle2: { fontSize: 14 },
 		},
 		shadows: [
 			"none", // 0
@@ -153,6 +164,7 @@ function createTheme() {
 				defaultProps: {
 					component: MuiAlert,
 					variant: "outlined",
+					severity: "none",
 					iconMapping: {
 						error: <ErrorIcon />,
 						info: <InfoIcon />,
@@ -231,6 +243,7 @@ function createTheme() {
 				defaultProps: {
 					component: MuiBadge,
 					color: "secondary",
+					slotProps: { badge: { component: MuiBadgeBadge } },
 				},
 			},
 			MuiBottomNavigation: { defaultProps: { component: Role.div } },
@@ -265,7 +278,12 @@ function createTheme() {
 					disableRipple: true, // ButtonGroup overrides Button's disableRipple so we need to set it here as well
 				},
 			},
-			MuiCard: { defaultProps: { component: MuiCard } },
+			MuiCard: {
+				defaultProps: {
+					component: MuiCard,
+					variant: "outlined",
+				},
+			},
 			MuiCardActionArea: {
 				defaultProps: {
 					component: MuiCardActionArea,
@@ -374,7 +392,7 @@ function createTheme() {
 			MuiMenuItem: { defaultProps: { component: Role.li } },
 			MuiMenuList: { defaultProps: { component: Role.ul } },
 			MuiMobileStepper: { defaultProps: { component: Role.div } },
-			MuiModal: { defaultProps: { component: Role.div } },
+			MuiModal: { defaultProps: { component: Role.div, container } },
 			MuiOutlinedInput: {
 				defaultProps: {
 					notched: false, // Removes masked border from Select
@@ -398,12 +416,19 @@ function createTheme() {
 				defaultProps: {
 					component: Role.div,
 					disableScrollLock: true,
+					// Popover passes down `container` prop to `Modal` https://github.com/mui/material-ui/blob/708ef10e874efa63d2e4972bd902befa1912f2dc/packages/mui-material/src/Popover/Popover.js#L389
+					container,
 					slots: {
 						paper: MuiPopoverPaperSlot,
 					},
 					slotProps: {
 						paper: { role: "dialog" },
 					},
+				},
+			},
+			MuiPopper: {
+				defaultProps: {
+					container,
 				},
 			},
 			MuiRadio: {
@@ -466,8 +491,24 @@ function createTheme() {
 			},
 			MuiSvgIcon: { defaultProps: { component: Role.svg } },
 			MuiSwipeableDrawer: { defaultProps: { component: Role.div } },
-			MuiTabs: { defaultProps: { component: Role.div } },
-			MuiTab: { defaultProps: { iconPosition: "start" } },
+			MuiTabs: {
+				defaultProps: {
+					component: MuiTabs,
+					allowScrollButtonsMobile: true,
+					slotProps: {
+						scrollButtons: {
+							"aria-hidden": true,
+						},
+						startScrollButtonIcon: {
+							component: ChevronLeftIcon,
+						},
+						endScrollButtonIcon: {
+							component: ChevronRightIcon,
+						},
+					},
+				},
+			},
+			MuiTab: { defaultProps: { component: MuiTab, iconPosition: "start" } },
 			MuiTable: { defaultProps: { component: withRenderProp(Role, "table") } },
 			MuiTableBody: {
 				defaultProps: {
@@ -532,7 +573,8 @@ function createTheme() {
 			},
 			MuiTypography: {
 				defaultProps: {
-					variant: "body2",
+					variant: "inherit",
+					variantMapping,
 					component: MuiTypography,
 				},
 				variants: [